refactor: rename components gallery module
