+++ src/components/gallery-page/ComponentsGalleryPanels.tsx
