+++ src/components/gallery-page/ComponentsPageClient.tsx
