+++ src/components/gallery-page/slug.ts
